+++ x/auth/legacy/v0_34/types.go
@@ -19,9 +19,8 @@ type (
 		SigVerifyCostSecp256k1 uint64 `json:"sig_verify_cost_secp256k1"`
 	}
 
-
 	GenesisState struct {
 		CollectedFees sdk.Coins `json:"collected_fees"`
-		Params        Params    `json:"params"`	
+		Params        Params    `json:"params"`
 	}
 )
