+++ x/distribution/legacy/v0_36/types.go
@@ -4,6 +4,7 @@ package v0_36
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	v034distr "github.com/cosmos/cosmos-sdk/x/distribution/legacy/v0_34"
 )
 
 // ----------------------------------------------------------------------------
@@ -17,93 +18,36 @@ const (
 type (
 	ValidatorAccumulatedCommission = sdk.DecCoins
 
-	ValidatorCurrentRewards struct {
-		Rewards sdk.DecCoins `json:"rewards"`
-		Period  uint64       `json:"period"`
-	}
-
-	ValidatorHistoricalRewards struct {
-		CumulativeRewardRatio sdk.DecCoins `json:"cumulative_reward_ratio"`
-		ReferenceCount        uint16       `json:"reference_count"`
-	}
-
-	DelegatorWithdrawInfo struct {
-		DelegatorAddress sdk.AccAddress `json:"delegator_address"`
-		WithdrawAddress  sdk.AccAddress `json:"withdraw_address"`
-	}
-
-	ValidatorOutstandingRewardsRecord struct {
-		ValidatorAddress   sdk.ValAddress `json:"validator_address"`
-		OutstandingRewards sdk.DecCoins   `json:"outstanding_rewards"`
-	}
-
-	ValidatorAccumulatedCommissionRecord struct {
-		ValidatorAddress sdk.ValAddress                 `json:"validator_address"`
-		Accumulated      ValidatorAccumulatedCommission `json:"accumulated"`
-	}
-
-	ValidatorHistoricalRewardsRecord struct {
-		ValidatorAddress sdk.ValAddress             `json:"validator_address"`
-		Period           uint64                     `json:"period"`
-		Rewards          ValidatorHistoricalRewards `json:"rewards"`
-	}
-
-	ValidatorCurrentRewardsRecord struct {
-		ValidatorAddress sdk.ValAddress          `json:"validator_address"`
-		Rewards          ValidatorCurrentRewards `json:"rewards"`
-	}
-
-	DelegatorStartingInfo struct {
-		PreviousPeriod uint64  `json:"previous_period"`
-		Stake          sdk.Dec `json:"stake"`
-		Height         uint64  `json:"creation_height"`
-	}
-
-	DelegatorStartingInfoRecord struct {
-		DelegatorAddress sdk.AccAddress        `json:"delegator_address"`
-		ValidatorAddress sdk.ValAddress        `json:"validator_address"`
-		StartingInfo     DelegatorStartingInfo `json:"starting_info"`
-	}
-
-	ValidatorSlashEvent struct {
-		ValidatorPeriod uint64  `json:"validator_period"`
-		Fraction        sdk.Dec `json:"fraction"`
-	}
-
 	ValidatorSlashEventRecord struct {
-		ValidatorAddress sdk.ValAddress      `json:"validator_address"`
-		Height           uint64              `json:"height"`
-		Period           uint64              `json:"period"`
-		Event            ValidatorSlashEvent `json:"validator_slash_event"`
-	}
-
-	FeePool struct {
-		CommunityPool sdk.DecCoins `json:"community_pool"`
+		ValidatorAddress sdk.ValAddress                `json:"validator_address"`
+		Height           uint64                        `json:"height"`
+		Period           uint64                        `json:"period"`
+		Event            v034distr.ValidatorSlashEvent `json:"validator_slash_event"`
 	}
 
 	GenesisState struct {
-		FeePool                         FeePool                                `json:"fee_pool"`
-		CommunityTax                    sdk.Dec                                `json:"community_tax"`
-		BaseProposerReward              sdk.Dec                                `json:"base_proposer_reward"`
-		BonusProposerReward             sdk.Dec                                `json:"bonus_proposer_reward"`
-		WithdrawAddrEnabled             bool                                   `json:"withdraw_addr_enabled"`
-		DelegatorWithdrawInfos          []DelegatorWithdrawInfo                `json:"delegator_withdraw_infos"`
-		PreviousProposer                sdk.ConsAddress                        `json:"previous_proposer"`
-		OutstandingRewards              []ValidatorOutstandingRewardsRecord    `json:"outstanding_rewards"`
-		ValidatorAccumulatedCommissions []ValidatorAccumulatedCommissionRecord `json:"validator_accumulated_commissions"`
-		ValidatorHistoricalRewards      []ValidatorHistoricalRewardsRecord     `json:"validator_historical_rewards"`
-		ValidatorCurrentRewards         []ValidatorCurrentRewardsRecord        `json:"validator_current_rewards"`
-		DelegatorStartingInfos          []DelegatorStartingInfoRecord          `json:"delegator_starting_infos"`
-		ValidatorSlashEvents            []ValidatorSlashEventRecord            `json:"validator_slash_events"`
+		FeePool                         v034distr.FeePool                                `json:"fee_pool"`
+		CommunityTax                    sdk.Dec                                          `json:"community_tax"`
+		BaseProposerReward              sdk.Dec                                          `json:"base_proposer_reward"`
+		BonusProposerReward             sdk.Dec                                          `json:"bonus_proposer_reward"`
+		WithdrawAddrEnabled             bool                                             `json:"withdraw_addr_enabled"`
+		DelegatorWithdrawInfos          []v034distr.DelegatorWithdrawInfo                `json:"delegator_withdraw_infos"`
+		PreviousProposer                sdk.ConsAddress                                  `json:"previous_proposer"`
+		OutstandingRewards              []v034distr.ValidatorOutstandingRewardsRecord    `json:"outstanding_rewards"`
+		ValidatorAccumulatedCommissions []v034distr.ValidatorAccumulatedCommissionRecord `json:"validator_accumulated_commissions"`
+		ValidatorHistoricalRewards      []v034distr.ValidatorHistoricalRewardsRecord     `json:"validator_historical_rewards"`
+		ValidatorCurrentRewards         []v034distr.ValidatorCurrentRewardsRecord        `json:"validator_current_rewards"`
+		DelegatorStartingInfos          []v034distr.DelegatorStartingInfoRecord          `json:"delegator_starting_infos"`
+		ValidatorSlashEvents            []ValidatorSlashEventRecord                      `json:"validator_slash_events"`
 	}
 )
 
 func NewGenesisState(
-	feePool FeePool, communityTax, baseProposerReward, bonusProposerReward sdk.Dec,
-	withdrawAddrEnabled bool, dwis []DelegatorWithdrawInfo, pp sdk.ConsAddress,
-	r []ValidatorOutstandingRewardsRecord, acc []ValidatorAccumulatedCommissionRecord,
-	historical []ValidatorHistoricalRewardsRecord, cur []ValidatorCurrentRewardsRecord,
-	dels []DelegatorStartingInfoRecord, slashes []ValidatorSlashEventRecord,
+	feePool v034distr.FeePool, communityTax, baseProposerReward, bonusProposerReward sdk.Dec,
+	withdrawAddrEnabled bool, dwis []v034distr.DelegatorWithdrawInfo, pp sdk.ConsAddress,
+	r []v034distr.ValidatorOutstandingRewardsRecord, acc []v034distr.ValidatorAccumulatedCommissionRecord,
+	historical []v034distr.ValidatorHistoricalRewardsRecord, cur []v034distr.ValidatorCurrentRewardsRecord,
+	dels []v034distr.DelegatorStartingInfoRecord, slashes []ValidatorSlashEventRecord,
 ) GenesisState {
 
 	return GenesisState{
