--- x/gov/legacy/v034/types.go
@@ -1,513 +1,0 @@
-package v034
-
-import (
-	"encoding/json"
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-// Keys
-const (
-	// ModuleName is the name of the module
-	ModuleName = "gov"
-
-	// StoreKey is the store key string for gov
-	StoreKey = ModuleName
-
-	// RouterKey is the message route for gov
-	RouterKey = ModuleName
-
-	// QuerierRoute is the querier route for gov
-	QuerierRoute = ModuleName
-
-	// DefaultParamspace default name for parameter store
-	DefaultParamspace = ModuleName
-)
-
-type GenesisState struct {
-	StartingProposalID uint64                `json:"starting_proposal_id"`
-	Deposits           []DepositWithMetadata `json:"deposits"`
-	Votes              []VoteWithMetadata    `json:"votes"`
-	Proposals          []Proposal            `json:"proposals"`
-	DepositParams      DepositParams         `json:"deposit_params"`
-	VotingParams       VotingParams          `json:"voting_params"`
-	TallyParams        TallyParams           `json:"tally_params"`
-}
-
-type DepositWithMetadata struct {
-	ProposalID uint64  `json:"proposal_id"`
-	Deposit    Deposit `json:"deposit"`
-}
-
-type VoteWithMetadata struct {
-	ProposalID uint64 `json:"proposal_id"`
-	Vote       Vote   `json:"vote"`
-}
-
-type Deposit struct {
-	ProposalID uint64         `json:"proposal_id"` //  proposalID of the proposal
-	Depositor  sdk.AccAddress `json:"depositor"`   //  Address of the depositor
-	Amount     sdk.Coins      `json:"amount"`      //  Deposit amount
-}
-
-type Deposits []Deposit
-
-type Vote struct {
-	ProposalID uint64         `json:"proposal_id"` //  proposalID of the proposal
-	Voter      sdk.AccAddress `json:"voter"`       //  address of the voter
-	Option     VoteOption     `json:"option"`      //  option from OptionSet chosen by the voter
-}
-
-type Votes []Vote
-
-// Param around deposits for governance
-type DepositParams struct {
-	MinDeposit       sdk.Coins     `json:"min_deposit,omitempty"`        //  Minimum deposit for a proposal to enter voting period.
-	MaxDepositPeriod time.Duration `json:"max_deposit_period,omitempty"` //  Maximum period for Atom holders to deposit on a proposal. Initial value: 2 months
-}
-
-type TallyParams struct {
-	Quorum    sdk.Dec `json:"quorum,omitempty"`    //  Minimum percentage of total stake needed to vote for a result to be considered valid
-	Threshold sdk.Dec `json:"threshold,omitempty"` //  Minimum proportion of Yes votes for proposal to pass. Initial value: 0.5
-	Veto      sdk.Dec `json:"veto,omitempty"`      //  Minimum value of Veto votes to Total votes ratio for proposal to be vetoed. Initial value: 1/3
-}
-
-type VotingParams struct {
-	VotingPeriod time.Duration `json:"voting_period,omitempty"` //  Length of the voting period.
-}
-
-type TallyResult struct {
-	Yes        sdk.Int `json:"yes"`
-	Abstain    sdk.Int `json:"abstain"`
-	No         sdk.Int `json:"no"`
-	NoWithVeto sdk.Int `json:"no_with_veto"`
-}
-
-// ----------------------------------------------------------------------------
-// ProposalStatus
-// ----------------------------------------------------------------------------
-
-// ProposalStatus is a type alias that represents a proposal status as a byte
-type ProposalStatus byte
-
-//nolint
-const (
-	StatusNil           ProposalStatus = 0x00
-	StatusDepositPeriod ProposalStatus = 0x01
-	StatusVotingPeriod  ProposalStatus = 0x02
-	StatusPassed        ProposalStatus = 0x03
-	StatusRejected      ProposalStatus = 0x04
-	StatusFailed        ProposalStatus = 0x05
-)
-
-// ProposalStatusToString turns a string into a ProposalStatus
-func ProposalStatusFromString(str string) (ProposalStatus, error) {
-	switch str {
-	case "DepositPeriod":
-		return StatusDepositPeriod, nil
-
-	case "VotingPeriod":
-		return StatusVotingPeriod, nil
-
-	case "Passed":
-		return StatusPassed, nil
-
-	case "Rejected":
-		return StatusRejected, nil
-
-	case "Failed":
-		return StatusFailed, nil
-
-	case "":
-		return StatusNil, nil
-
-	default:
-		return ProposalStatus(0xff), fmt.Errorf("'%s' is not a valid proposal status", str)
-	}
-}
-
-// Marshal needed for protobuf compatibility
-func (status ProposalStatus) Marshal() ([]byte, error) {
-	return []byte{byte(status)}, nil
-}
-
-// Unmarshal needed for protobuf compatibility
-func (status *ProposalStatus) Unmarshal(data []byte) error {
-	*status = ProposalStatus(data[0])
-	return nil
-}
-
-// Marshals to JSON using string
-func (status ProposalStatus) MarshalJSON() ([]byte, error) {
-	return json.Marshal(status.String())
-}
-
-// Unmarshals from JSON assuming Bech32 encoding
-func (status *ProposalStatus) UnmarshalJSON(data []byte) error {
-	var s string
-	err := json.Unmarshal(data, &s)
-	if err != nil {
-		return err
-	}
-
-	bz2, err := ProposalStatusFromString(s)
-	if err != nil {
-		return err
-	}
-
-	*status = bz2
-	return nil
-}
-
-// String implements the Stringer interface.
-func (status ProposalStatus) String() string {
-	switch status {
-	case StatusDepositPeriod:
-		return "DepositPeriod"
-
-	case StatusVotingPeriod:
-		return "VotingPeriod"
-
-	case StatusPassed:
-		return "Passed"
-
-	case StatusRejected:
-		return "Rejected"
-
-	case StatusFailed:
-		return "Failed"
-
-	default:
-		return ""
-	}
-}
-
-// Format implements the fmt.Formatter interface.
-// nolint: errcheck
-func (status ProposalStatus) Format(s fmt.State, verb rune) {
-	switch verb {
-	case 's':
-		s.Write([]byte(status.String()))
-	default:
-		// TODO: Do this conversion more directly
-		s.Write([]byte(fmt.Sprintf("%v", byte(status))))
-	}
-}
-
-// ----------------------------------------------------------------------------
-// VoteOption
-// ----------------------------------------------------------------------------
-
-type VoteOption byte
-
-// Vote options
-const (
-	OptionEmpty      VoteOption = 0x00
-	OptionYes        VoteOption = 0x01
-	OptionAbstain    VoteOption = 0x02
-	OptionNo         VoteOption = 0x03
-	OptionNoWithVeto VoteOption = 0x04
-)
-
-// VoteOptionFromString returns a VoteOption from a string. It returns an error
-// if the string is invalid.
-func VoteOptionFromString(str string) (VoteOption, error) {
-	switch str {
-	case "Yes":
-		return OptionYes, nil
-
-	case "Abstain":
-		return OptionAbstain, nil
-
-	case "No":
-		return OptionNo, nil
-
-	case "NoWithVeto":
-		return OptionNoWithVeto, nil
-
-	default:
-		return VoteOption(0xff), fmt.Errorf("'%s' is not a valid vote option", str)
-	}
-}
-
-// Marshal needed for protobuf compatibility.
-func (vo VoteOption) Marshal() ([]byte, error) {
-	return []byte{byte(vo)}, nil
-}
-
-// Unmarshal needed for protobuf compatibility.
-func (vo *VoteOption) Unmarshal(data []byte) error {
-	*vo = VoteOption(data[0])
-	return nil
-}
-
-// Marshals to JSON using string.
-func (vo VoteOption) MarshalJSON() ([]byte, error) {
-	return json.Marshal(vo.String())
-}
-
-// UnmarshalJSON decodes from JSON assuming Bech32 encoding.
-func (vo *VoteOption) UnmarshalJSON(data []byte) error {
-	var s string
-	err := json.Unmarshal(data, &s)
-	if err != nil {
-		return err
-	}
-
-	bz2, err := VoteOptionFromString(s)
-	if err != nil {
-		return err
-	}
-
-	*vo = bz2
-	return nil
-}
-
-// String implements the Stringer interface.
-func (vo VoteOption) String() string {
-	switch vo {
-	case OptionYes:
-		return "Yes"
-	case OptionAbstain:
-		return "Abstain"
-	case OptionNo:
-		return "No"
-	case OptionNoWithVeto:
-		return "NoWithVeto"
-	default:
-		return ""
-	}
-}
-
-// Format implements the fmt.Formatter interface.
-// nolint: errcheck
-func (vo VoteOption) Format(s fmt.State, verb rune) {
-	switch verb {
-	case 's':
-		s.Write([]byte(vo.String()))
-	default:
-		s.Write([]byte(fmt.Sprintf("%v", byte(vo))))
-	}
-}
-
-// ----------------------------------------------------------------------------
-// Tally
-// ----------------------------------------------------------------------------
-
-// Equals returns if two proposals are equal.
-func (tr TallyResult) Equals(comp TallyResult) bool {
-	return tr.Yes.Equal(comp.Yes) &&
-		tr.Abstain.Equal(comp.Abstain) &&
-		tr.No.Equal(comp.No) &&
-		tr.NoWithVeto.Equal(comp.NoWithVeto)
-}
-
-func (tr TallyResult) String() string {
-	return fmt.Sprintf(`Tally Result:
-  Yes:        %s
-  Abstain:    %s
-  No:         %s
-  NoWithVeto: %s`, tr.Yes, tr.Abstain, tr.No, tr.NoWithVeto)
-}
-
-// ----------------------------------------------------------------------------
-// Proposal
-// ----------------------------------------------------------------------------
-
-type Proposal struct {
-	ProposalContent `json:"proposal_content"` // Proposal content interface
-
-	ProposalID uint64 `json:"proposal_id"` //  ID of the proposal
-
-	Status           ProposalStatus `json:"proposal_status"`    //  Status of the Proposal {Pending, Active, Passed, Rejected}
-	FinalTallyResult TallyResult    `json:"final_tally_result"` //  Result of Tallys
-
-	SubmitTime     time.Time `json:"submit_time"`      //  Time of the block where TxGovSubmitProposal was included
-	DepositEndTime time.Time `json:"deposit_end_time"` // Time that the Proposal would expire if deposit amount isn't met
-	TotalDeposit   sdk.Coins `json:"total_deposit"`    //  Current deposit on this proposal. Initial value is set at InitialDeposit
-
-	VotingStartTime time.Time `json:"voting_start_time"` //  Time of the block where MinDeposit was reached. -1 if MinDeposit is not reached
-	VotingEndTime   time.Time `json:"voting_end_time"`   // Time that the VotingPeriod for this proposal will end and votes will be tallied
-}
-
-// nolint
-func (p Proposal) String() string {
-	return fmt.Sprintf(`Proposal %d:
-  Title:              %s
-  Type:               %s
-  Status:             %s
-  Submit Time:        %s
-  Deposit End Time:   %s
-  Total Deposit:      %s
-  Voting Start Time:  %s
-  Voting End Time:    %s
-  Description:        %s`,
-		p.ProposalID, p.GetTitle(), p.ProposalType(),
-		p.Status, p.SubmitTime, p.DepositEndTime,
-		p.TotalDeposit, p.VotingStartTime, p.VotingEndTime, p.GetDescription(),
-	)
-}
-
-// ProposalContent is an interface that has title, description, and proposaltype
-// that the governance module can use to identify them and generate human readable messages
-// ProposalContent can have additional fields, which will handled by ProposalHandlers
-// via type assertion, e.g. parameter change amount in ParameterChangeProposal
-type ProposalContent interface {
-	GetTitle() string
-	GetDescription() string
-	ProposalType() ProposalKind
-}
-
-// Proposals is an array of proposal
-type Proposals []Proposal
-
-// nolint
-func (p Proposals) String() string {
-	out := "ID - (Status) [Type] Title\n"
-	for _, prop := range p {
-		out += fmt.Sprintf("%d - (%s) [%s] %s\n",
-			prop.ProposalID, prop.Status,
-			prop.ProposalType(), prop.GetTitle())
-	}
-	return strings.TrimSpace(out)
-}
-
-// Text Proposals
-type TextProposal struct {
-	Title       string `json:"title"`       //  Title of the proposal
-	Description string `json:"description"` //  Description of the proposal
-}
-
-func NewTextProposal(title, description string) TextProposal {
-	return TextProposal{
-		Title:       title,
-		Description: description,
-	}
-}
-
-// Implements Proposal Interface
-var _ ProposalContent = TextProposal{}
-
-// nolint
-func (tp TextProposal) GetTitle() string           { return tp.Title }
-func (tp TextProposal) GetDescription() string     { return tp.Description }
-func (tp TextProposal) ProposalType() ProposalKind { return ProposalTypeText }
-
-// Software Upgrade Proposals
-type SoftwareUpgradeProposal struct {
-	TextProposal
-}
-
-func NewSoftwareUpgradeProposal(title, description string) SoftwareUpgradeProposal {
-	return SoftwareUpgradeProposal{
-		TextProposal: NewTextProposal(title, description),
-	}
-}
-
-// Implements Proposal Interface
-var _ ProposalContent = SoftwareUpgradeProposal{}
-
-// nolint
-func (sup SoftwareUpgradeProposal) ProposalType() ProposalKind { return ProposalTypeSoftwareUpgrade }
-
-// ProposalQueue
-type ProposalQueue []uint64
-
-// ProposalKind
-
-// Type that represents Proposal Type as a byte
-type ProposalKind byte
-
-//nolint
-const (
-	ProposalTypeNil             ProposalKind = 0x00
-	ProposalTypeText            ProposalKind = 0x01
-	ProposalTypeParameterChange ProposalKind = 0x02
-	ProposalTypeSoftwareUpgrade ProposalKind = 0x03
-)
-
-// String to proposalType byte. Returns 0xff if invalid.
-func ProposalTypeFromString(str string) (ProposalKind, error) {
-	switch str {
-	case "Text":
-		return ProposalTypeText, nil
-	case "ParameterChange":
-		return ProposalTypeParameterChange, nil
-	case "SoftwareUpgrade":
-		return ProposalTypeSoftwareUpgrade, nil
-	default:
-		return ProposalKind(0xff), fmt.Errorf("'%s' is not a valid proposal type", str)
-	}
-}
-
-// Marshal needed for protobuf compatibility
-func (pt ProposalKind) Marshal() ([]byte, error) {
-	return []byte{byte(pt)}, nil
-}
-
-// Unmarshal needed for protobuf compatibility
-func (pt *ProposalKind) Unmarshal(data []byte) error {
-	*pt = ProposalKind(data[0])
-	return nil
-}
-
-// Marshals to JSON using string
-func (pt ProposalKind) MarshalJSON() ([]byte, error) {
-	return json.Marshal(pt.String())
-}
-
-// Unmarshals from JSON assuming Bech32 encoding
-func (pt *ProposalKind) UnmarshalJSON(data []byte) error {
-	var s string
-	err := json.Unmarshal(data, &s)
-	if err != nil {
-		return err
-	}
-
-	bz2, err := ProposalTypeFromString(s)
-	if err != nil {
-		return err
-	}
-	*pt = bz2
-	return nil
-}
-
-// Turns VoteOption byte to String
-func (pt ProposalKind) String() string {
-	switch pt {
-	case ProposalTypeText:
-		return "Text"
-	case ProposalTypeParameterChange:
-		return "ParameterChange"
-	case ProposalTypeSoftwareUpgrade:
-		return "SoftwareUpgrade"
-	default:
-		return ""
-	}
-}
-
-// For Printf / Sprintf, returns bech32 when using %s
-// nolint: errcheck
-func (pt ProposalKind) Format(s fmt.State, verb rune) {
-	switch verb {
-	case 's':
-		s.Write([]byte(pt.String()))
-	default:
-		// TODO: Do this conversion more directly
-		s.Write([]byte(fmt.Sprintf("%v", byte(pt))))
-	}
-}
-
-// ----------------------------------------------------------------------------
-// Codec
-// ----------------------------------------------------------------------------
-
-func RegisterCodec(cdc *codec.Codec) {
-	cdc.RegisterInterface((*ProposalContent)(nil), nil)
-	cdc.RegisterConcrete(TextProposal{}, "gov/TextProposal", nil)
-	cdc.RegisterConcrete(SoftwareUpgradeProposal{}, "gov/SoftwareUpgradeProposal", nil)
-}
